rename component directory to lower-case
