+++ src/js/component/DirtyStates.js
