+++ src/js/component/Polling.js
