+++ src/js/component/PrefetchManager.js
