+++ src/js/component/UpdateQueryString.js
