+++ src/js/component/index.js
